添加 WebSocket 和 SSL

+++ sylar/stream.cc
@@ -5,9 +5,9 @@ namespace sylar {
 
 int Stream::readFixSize(void* buffer, size_t length) {
     size_t offset = 0;
-    size_t left = length;
+    int64_t left = length;
     while (left > 0) {
-        size_t len = read((char*)buffer + offset, left);
+        int64_t len = read((char*)buffer + offset, left);
         if (len <= 0) {
             return len;
         }
@@ -18,9 +18,9 @@ int Stream::readFixSize(void* buffer, size_t length) {
 }
 
 int Stream::readFixSize(ByteArray::ptr ba, size_t length) {
-    size_t left = length;
+    int64_t left = length;
     while (left > 0) {
-        size_t len = read(ba, left);  // ba 有 position 属性
+        int64_t len = read(ba, left);  // ba 有 position 属性
         if (len <= 0) {
             return len;
         }
@@ -31,9 +31,9 @@ int Stream::readFixSize(ByteArray::ptr ba, size_t length) {
 
 int Stream::writeFixSize(const void* buffer, size_t length) {
     size_t offset = 0;
-    size_t left = length;
+    int64_t left = length;
     while (left > 0) {
-        size_t len = write((const char*)buffer + offset, left);
+        int64_t len = write((const char*)buffer + offset, left);
         if (len <= 0) {
             return len;
         }
@@ -44,9 +44,9 @@ int Stream::writeFixSize(const void* buffer, size_t length) {
 }
 
 int Stream::writeFixSize(ByteArray::ptr ba, size_t length) {
-    size_t left = length;
+    int64_t left = length;
     while (left > 0) {
-        size_t len = write(ba, left);
+        int64_t len = write(ba, left);
         if (len <= 0) {
             return len;
         }
